requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
django-filter>=23.3
psycopg2-binary>=2.9.7
python-decouple>=3.8
Pillow>=9.0.0
celery>=5.3.0
redis>=4.5.4
gunicorn>=21.2.0
djoser>=2.1.0
phonenumberslite>=8.13.13
django-phonenumber-field>=7.0.2
requests>=2.31.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4